Add flush to the print statements
We will end up getting rid of the print statements in
the future, but it helps for simple troubleshooting in
the beginning. Adding the flush flag helps with
troubleshooting in docker since stdout isn't flushed
all the time. It'll hurt performance a bit, but not a
big deal at this point.